requirements.txt
SpeechRecognition
pydub
pyaudio
openai-whisper
numpy<2
scipy
langdetect
flask
pytest
